Added UI elements, Settings screen, MainScreen Controller, Settings controller,extra debugging, partly cleaned up code.
NumberWizard.cs: added getters/setters, logging code for debugging
purposes (will be cleaned up). Separated initial difficulty from numbers
of guesses. Removed default initialisation of min and max values (left
that to the settings-script).
NumberSettings.cs: completely new. Allows for the difficulty (number of
guesses)to be adjusted, as well as the min and max vars.
MainScreen.cs handles the subtitle text if the settings have been
changed. Might be a messy way to do it but works for now
LevelManager.cs: added code so the settings will be saved.

Next steps are cleaning up the code.